A good working version, with AudioFocusChangeListener and UtteranceProgressListener.
Successfully suppress music and resume it after speech is done; successfully stop app speech
when the map starts to talk.

The only problem is when map is in the middle of talking, and this speech comes up, it would
still gain audio focus and overlapping with map voices, with ducked music in the background.